Add files via upload
This is a UTC project with image data that has been set up and visualised succesfully. The images can be transformed and are easy to work with. Our data has been already split to train, test and valid form so we do not need to do splitting before the preprocessing and modeling of our data set.